Go-Green is a project to analyze the greenery in certain area.And to provide detail analysis of greenery requirement in that area.
The aim is to detect number of trees, calculate amount of oxygen released by these trees in that particular area. On the basis of number of people (population in that area) provided as an input by the user, we are doing the analysis of amount of Carbon dioxide exhaled by these people as well as Carbon dioxide generated due to the automobiles running in that locality. On the basis of this data collection we can analyze the amount of Oxygen required in that locality or it is present in ample amount.